remove outdated tf code
